requirements.txt
Flask==0.11.1
Pillow==3.3.0
Peewee==2.8.1